+++ Signal/Pulsar/CovarianceMatrixEngineCUDA.cu
@@ -11,7 +11,7 @@
 
 void computeCovarianceMatrixCUDAEngine(float* d_resultVector, unsigned int resultByteOffset,
 		const float* h_amps, float* d_amps, unsigned int ampsLength,
-		 const unsigned int* h_hits, float* d_hits, unsigned int hitsLength,
+		 const unsigned int* h_hits, unsigned int* d_hits, unsigned int hitsLength,
 		 unsigned int stokesLength, double scaleFactor, unsigned int blockDim2D)
 {
