+++ Signal/Pulsar/dsp/CovarianceMatrixCUDA.h
@@ -14,7 +14,7 @@
 
 __global__ void outerProductKernel(float* resultMatrix, float* vec, int vectorLength);
 
-__global__ void meanStokesKernel(float* d_amps, unsigned int ampsLength, float* d_hits, unsigned int stokesLength);
+__global__ void meanStokesKernel(float* d_amps, unsigned int ampsLength, unsigned int* d_hits, unsigned int stokesLength);
 
 
 __global__ void applyScale(float* amps, unsigned int ampsLength, double scaleFactor);
